test: compare older and new
* compare order and new
